* test: adds test for bugs in bruteforce dispatcher
* undone mistakenly made changes in test_cython_bruteforce_dispatcher

Fix bruteforce bugs (#135)

* fix(cythonic-bruteforce-dispatcher): fixes failing test test_append_due_to_timewindow (cython only)

there was an indexing error in is_timewindow_violated_dueto_insertion and int instead of bool in signature. ~ operator was being applied on the return value, instead of !

* fix(bruteforce-dispatcher): fixes the bugs in python

* fix(bruteforce-dispatcher): fixes the same thing in cython

* fix: fixes test_equivalence_simulator_cython_and_python_bruteforce_dispatcher

that had different request rates between python and cython

* fix(test_pickling\): ground truth changed as a result of float -> double

+++ thesimulator/data_structures_cython/data_structures.pyx
@@ -104,7 +104,7 @@ cdef class TransportationRequest(Request):
     def __init__(
             self,
             int request_id,
-            float creation_timestamp,
+            double creation_timestamp,
             origin,
             destination,
             pickup_timewindow_min,
@@ -330,7 +330,7 @@ cdef class InternalRequest(Request):
        `cdef InternalRequest from_c_r2loc` and `cdef InternalRequest from_c_int` for details.
     """
     def __init__(
-            self, int request_id, float creation_timestamp,
+            self, int request_id, double creation_timestamp,
             location
     ):
         """
